web/requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Caching==1.10.1
python-slugify==8.0.1
gunicorn==20.1.0
pytest==6.2.5
pytest-cov==2.12.1
black==21.12b0
flake8==4.0.1
